account type guard move to utils
