+++ src/interfaces/invoice.ts
@@ -1,6 +1,6 @@
 import { IPayment, isIPayment } from '@/interfaces/payment';
-import { EventType } from '@/interfaces/account';
-import { isEventType } from '@/libs/utils/type_guard/account';
+import { EventType } from '@/constants/account';
+import { isEventType } from '@/lib/utils/type_guard/account';
 
 export interface IInvoice {
   invoiceId: string;
